[41기 이민규] Assignment 2
