+++ minkyu/.env.sample
@@ -1,0 +1,10 @@
+DATABASE_URL="mysql://username:urpassword@ipadress/database_name"
+
+DATABASE_URL = DATABASE_URL
+
+TYPEORM_CONNECTION = CONNECTION
+TYPEORM_HOST = HOST
+TYPEORM_USERNAME = USERNAME 
+TYPEORM_PASSWORD = PASSWORD 
+TYPEORM_PORT = PORT
+TYPEORM_LOGGING = LOGGING
