+++ minkyu/package.json
@@ -2,9 +2,9 @@
   "name": "minkyu",
   "version": "1.0.0",
   "description": "",
-  "main": "express_init.js",
+  "main": "app.js",
   "scripts": {
-    "start": "nodemon express_init.js",
+    "start": "nodemon app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -12,6 +12,7 @@
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
+    "dbmate": "^1.0.3",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "morgan": "^1.10.0",
